claude-code/skills/video-to-summary-v2/requirements.txt
# Video to Summary V2 - Python Dependencies

# YAML configuration
pyyaml>=6.0

# HTTP requests
requests>=2.28.0

# HTML parsing
beautifulsoup4>=4.12.0

# Note: yt-dlp, ffmpeg, and whisper are installed separately
# See install.sh for system dependencies
